Add checkbox to enable saving previous extras
I don't forsee any issues with this, it doesn't do any special checking
for if the manifest changes in the future but it shouldn't matter, if
the entries don't exist it simply won't check anything.

I also changed all occurences of FindSteamSkinDir() to SteamSkinPath
since that String already existed. I should have included that in the
previous commit but only noticed as I was adding in the save extras
code.

And lastly I moved the reading of the manifest out of the for loop,
certainly doesn't need to be redone each iteration of the loop, again
this could have been including in previous commit but I didn't realize
at the time.